% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phylogenomics_Workshop-package.R
\docType{package}
\name{Phylogenomics_Workshop-package}
\alias{Phylogenomics_Workshop}
\alias{Phylogenomics_Workshop-package}
\title{Phylogenomics_Workshop: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LPDagallier/Phylogenomics_Workshop}
  \item Report bugs at \url{https://github.com/LPDagallier/Phylogenomics_Workshop/issues}
}

}
\author{
\strong{Maintainer}: Leo-Paul Dagallier \email{leopauldagallier@gmail.com} (\href{https://orcid.org/0000-0002-3270-1544}{ORCID}) [copyright holder]

}
\keyword{internal}
